Modernize dateField syntax
